[generators] first classes to use a new parsing system
Classes responsible to parse wiki content are not responsible anymore to generate output content.
They must call a generator to indicate data to use to generate content, and
generators are responsible to generate output content.

If we want to generate a wiki content to a new text syntax, we don't have to
rewrite all wiki parsers.